pyproject.toml
[project]
name = "discord-ai-news"
version = "0.1.0"
description = "Automatic send news to Discord for various topics"
authors = [
    {name = "Timothe Duquesne",email = "timothe.duquesne@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-genai (>=1.53.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
